% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{build_ui}
\alias{build_ui}
\title{Build the Shiny application UI}
\usage{
build_ui()
}
\value{
A \code{\link[shiny:fluidPage]{shiny::fluidPage()}} UI object.
}
\description{
Build the Shiny application UI
}
\keyword{internal}
