% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph.R
\name{mod_graph_server}
\alias{mod_graph_server}
\title{Server for the hierarchy graph module}
\usage{
mod_graph_server(id, merged_rv, cache_dir, highlight_terms_rv = NULL)
}
\arguments{
\item{id}{Shiny module namespace id.}

\item{merged_rv}{ReactiveVal with merged glossary data.}

\item{cache_dir}{Path to app cache directory.}

\item{highlight_terms_rv}{Optional \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} used to push graph
selection highlights to other modules.}
}
\description{
Server for the hierarchy graph module
}
\keyword{internal}
