% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph.R
\name{mod_graph_ui}
\alias{mod_graph_ui}
\title{UI for the hierarchy graph module}
\usage{
mod_graph_ui(id)
}
\arguments{
\item{id}{Shiny module namespace id.}
}
\value{
UI tags.
}
\description{
UI for the hierarchy graph module
}
\keyword{internal}
