% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_server}
\alias{mod_table_server}
\title{Server for the glossary comparison table module}
\usage{
mod_table_server(id, merged_rv, cache_dir, highlight_terms_rv = NULL)
}
\arguments{
\item{id}{Shiny module namespace id.}

\item{merged_rv}{A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} containing the merged glossary
\code{\link[tibble:tibble]{tibble::tibble()}} (output of \code{\link[=merge_glossaries]{merge_glossaries()}}).}

\item{cache_dir}{Path to the cache directory.}

\item{highlight_terms_rv}{Optional \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} containing a
character vector of terms to highlight in the table.}
}
\description{
Server for the glossary comparison table module
}
\keyword{internal}
