% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_ui}
\alias{mod_table_ui}
\title{UI for the glossary comparison table module}
\usage{
mod_table_ui(id)
}
\arguments{
\item{id}{Shiny module namespace id.}
}
\value{
A \code{reactable} output wrapped in a \code{div}.
}
\description{
UI for the glossary comparison table module
}
\keyword{internal}
