% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_ipcc.R
\name{mod_update_ipcc_server}
\alias{mod_update_ipcc_server}
\title{Server for the IPCC update module}
\usage{
mod_update_ipcc_server(id, cache_dir, merged_rv)
}
\arguments{
\item{id}{Shiny module namespace id.}

\item{cache_dir}{Path to the user cache directory.}

\item{merged_rv}{A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} to update after a successful scrape.}
}
\description{
Server for the IPCC update module
}
\keyword{internal}
