% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_ipcc.R
\name{mod_update_ipcc_ui}
\alias{mod_update_ipcc_ui}
\title{UI for the IPCC update module}
\usage{
mod_update_ipcc_ui(id)
}
\arguments{
\item{id}{Shiny module namespace id.}
}
\value{
A \code{tagList} with the update button, progress placeholder, and status
text.
}
\description{
UI for the IPCC update module
}
\keyword{internal}
