% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise_term}
\alias{normalise_term}
\title{Normalise a glossary term for matching}
\usage{
normalise_term(term)
}
\arguments{
\item{term}{Character vector of term names.}
}
\value{
Normalised character vector.
}
\description{
Lowercases, strips non-alphanumeric characters (except spaces), and trims
whitespace.  Used to match IPBES concept names against IPCC term names.
}
\keyword{internal}
