% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity_bar_html}
\alias{similarity_bar_html}
\title{Build a coloured similarity bar HTML snippet}
\usage{
similarity_bar_html(score)
}
\arguments{
\item{score}{Numeric 0-1 (or NA).}
}
\value{
HTML string.
}
\description{
Build a coloured similarity bar HTML snippet
}
\keyword{internal}
