% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity_triplet_html}
\alias{similarity_triplet_html}
\title{Build a three-line similarity HTML snippet}
\usage{
similarity_triplet_html(sim_within_ipbes, sim_within_ipcc, sim_between_all)
}
\arguments{
\item{sim_within_ipbes}{Numeric 0-1 or \code{NA}.}

\item{sim_within_ipcc}{Numeric 0-1 or \code{NA}.}

\item{sim_between_all}{Numeric 0-1 or \code{NA}.}
}
\value{
HTML string.
}
\description{
Build a three-line similarity HTML snippet
}
\keyword{internal}
