% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{STOPWORDS}
\alias{STOPWORDS}
\title{Common English stopwords}
\format{
An object of class \code{character} of length 71.
}
\usage{
STOPWORDS
}
\description{
A character vector of ~60 common English stopwords used when tokenising
definition text for similarity computation.
}
\keyword{internal}
