% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ipcc.R
\name{summarise_ipcc}
\alias{summarise_ipcc}
\title{Summarise IPCC glossary to one row per term}
\usage{
summarise_ipcc(ipcc_df)
}
\arguments{
\item{ipcc_df}{Output of \code{\link[=load_ipcc]{load_ipcc()}}.}
}
\value{
A \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{term}, \code{n_reports},
\code{summary_definition}, \code{ipcc_data} (list-column of report x definition
tibbles).
}
\description{
Summarise IPCC glossary to one row per term
}
